Revert "Use auto_now_add for created and auto_now for updated for Inode."
This reverts commit d5e62193c2236fcd0a7ec300d5ffe209124bda34.

django timezone.now will save timezone-aware datetime object into db
when `USE_TZ = True` (which is default) in the settings.py

Besides, use `auto_now` and `auto_now_add` will make this field cannot
be overrided in the future.

We have to save the `created` and `updated` with the ctime and mtime of
the inode object in the file system in the future, so we should use
`default` not `auto_now` and `auto_now_add`.

ref:
https://docs.djangoproject.com/en/1.10/ref/models/fields/#django.db.models.DateField.auto_now_add